HUDGE UPDATE : Introducing the CLSSObject and CLSSFacility in order to simplify the shared properties between must of the CLASS object : Almost all the CLASS Object will inherit of the CLSSObject class, which actually only include the log file.(Except class ZAI IsotopicVector) All facility in CLASS now inherit of the CLSSFacility, which are a CLSSObject and a TNamed (from ROOT/CERN library). In the CLSSFacility, one can found : fId, fParc, fCreationTime, fLifeTime as private variable (the daughter class can not access directly to this member, one need to use the setter and getter - SetId(int) by ex-) fIsStarted, fShutDown, fEndOfCycle, fInternalTime, fInCycleTime, fCycleTime, fInsideIV as protected member (the daughter class can directly modify this variable - also accessible via setter/getter) this repartition between private and protected could change in the future, fCycleTime replace fCoolingTime and fFabricationTime in respectively Pool and FabricationPlant
Those variable are shared by almost all the facility, it is then more convenient to regroup them. 
For all facility the fInsideIV represent the total amount of fuel in the facility, in the case of the Reactor, this is also the fuel inside the reactor (it replace the old fReactorIV)

If it is needed to add any variable to all CLASS Object (except ZAI and IsotopicVector), one should use the CLSSObject, for the facility, use CLSSFacility.

+ All ClassDef of the daughter facilities change from 1 to 2 (because of the inheritance), the ClassDef of the CLSSObject has been set to 0, and 1 for CLSSFacility (as explain in rootcin doc, 0 mean no I/O, 1 is the first level, and should increase by 1 at each inheritance...)

Resolve the small glitch (not really a bug) which cause the warning at the execution of a scenario, and at the reading of the output in ROOT, using #ifndef __CINT__ // #endif  to exclude some part of the variable at the compilation of the root library...


BaM


git-svn-id: svn+ssh://svn.in2p3.fr/class@116 0e7d625b-0364-4367-a6be-d5be4a48d228